requirements.txt
google-generativeai>=0.8.0
google-api-python-client>=2.150.0
youtube-transcript-api>=0.6.3
edge-tts>=6.1.0
pydub>=0.25.1
pyyaml>=6.0.2
mutagen>=1.47.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
jinja2>=3.1.4
aiofiles>=24.1.0
python-dotenv>=1.0.1
